django/requirements.txt
Django
djangorestframework
psycopg2-binary
PyPDF2
